Moved plugins around
