+++ src/components/Plugin/DocViewer.tsx
@@ -1,8 +1,8 @@
-import { Plugin } from '../plugin';
+import { Plugin } from '../../types/Plugin';
 import { ReactNode, FC, useState, useEffect } from 'react';
 import DocViewer, { IDocument } from '@cyntler/react-doc-viewer';
-import { S3Object } from '../../../types/S3Object';
-import { useS3Context } from '../../../contexts/s3-context';
+import { S3Object } from '../../types/S3Object';
+import { useS3Context } from '../../contexts/s3-context';
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
 import { GetObjectCommand } from '@aws-sdk/client-s3';
