+++ src/components/S3Viewer/S3Viewer.tsx
@@ -3,7 +3,7 @@ import { FileMain } from './FileMain';
 import { S3Client } from '@aws-sdk/client-s3';
 import { S3Provider } from '../../contexts/s3-context';
 import { PluginManagerProvider } from '../../context/plugins.context';
-import { Plugin } from '../../utils/plugin/plugin';
+import { Plugin } from './../../types/Plugin';
 
 interface S3ViewerProps {
   client: S3Client;
