+++ src/types/Plugin.ts
@@ -1,4 +1,20 @@
-import { Plugin } from './plugin';
+import { ReactNode } from 'react';
+import { S3Object } from './S3Object';
+
+/**
+ * Definition of a Plugin. Represents the expected form of all plugins that
+ * are under the PluginManager
+ */
+export interface Plugin {
+  /** The name of the plugin in a human readable format */
+  name: string;
+  /** Description of the plugin in a human readable format */
+  description: string;
+  /** File extensions the plugin is associated with */
+  fileExtensions: string[];
+  /** Get the view to show up in the modal */
+  getView(object: S3Object): ReactNode;
+}
 
 export class PluginManager {
   /**
@@ -25,7 +41,6 @@ export class PluginManager {
         if (!this.pluginMapping.has(fileExtension)) {
           this.pluginMapping.set(fileExtension, [plugin]);
         } else {
-          console.log(this.pluginMapping);
           this.pluginMapping.get(fileExtension)!.push(plugin);
         }
       }
